+++ src/main/java/rasetech/student/management/StudentManagementApplication.java
@@ -1,35 +1,21 @@
 package rasetech.student.management;
 
 import java.util.List;
-import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import rasetech.student.management.data.Student;
+import rasetech.student.management.data.StudentCourses;
+import rasetech.student.management.repository.StudentRepository;
 
 @SpringBootApplication
-@RestController
 public class StudentManagementApplication {
-
-  @Autowired
-  private StudentRepository repository;
 
   public static void main(String[] args) {
     SpringApplication.run(StudentManagementApplication.class, args);
   }
 
-  @GetMapping("/student")
-  public List<Student> getStudentList() {
-    return repository.search();
-  }
 
-  @GetMapping("/student_coursesList")
-  public List<StudentCourses> getStudentCoursesList() {
-    return repository.searchStudentCourses();
-  }
 }
