--- src/main/java/rasetech/student/management/StudentRepository.java
@@ -1,16 +1,0 @@
-package rasetech.student.management;
-
-import java.util.List;
-import org.apache.ibatis.annotations.Delete;
-import org.apache.ibatis.annotations.Insert;
-import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Select;
-import org.apache.ibatis.annotations.Update;
-
-@Mapper
-public interface StudentRepository {
-@Select("SELECT * FROM students")
-  List<Student> search();
-@Select("SELECT * FROM student_courses")
-  List<StudentCourses> searchStudentCourses();
-}
