+++ src/main/java/rasetech/student/management/data/Student.java
@@ -1,6 +1,5 @@
-package rasetech.student.management;
+package rasetech.student.management.data;
 
-import java.time.LocalDateTime;
 import lombok.Getter;
 import lombok.Setter;
 @Getter
